Uhit tests for LightSearch Android. Logic tests. Part 1

+++ Standard Edition/LightSearch Android/LightSearch/app/src/main/java/ru/viise/lightsearch/data/SoftCheckRecordDefaultImpl.java
@@ -43,7 +43,7 @@ public class SoftCheckRecordDefaultImpl implements SoftCheckRecord {
         this.amountUnit = amountUnit;
         this.subdivisions = subdivisions;
 
-        int tempMaxAmount = 0;
+        float tempMaxAmount = 0;
         for(Subdivision subdivision : subdivisions.collection())
             tempMaxAmount += subdivision.productAmount();
